Add linter to lint ebuilds (#11)
Adds a command named 'elint' to lint ebuilds. This mainly ensures that
they have a few required fields (`DESCRIPTION` and `LICENSE` right now)
and that their `Manifest` files are valid. This is helpful for ensuring
that we don't accidentally ever have broken ebuilds in this repository.

Includes a Docker image for it to use so that the linter can be ran on
any operating system (`ebuild` only exists on Gentoo). We use the stage3
docker image and run `emerge-webrsync` on top of it to keep lint times
low.

Includes a Github action to build and push that Docker image (mostly
meant to keep it updated periodically, since you'd need to push it up
manually to get it working if it truely broke) as well as an action to
run `elint` on all PRs and pushes to `main`.